fix fan double click add 1% (#989)

+++ TFT/src/User/API/FanControl.h
@@ -9,7 +9,7 @@ extern const char* fanCmd[];
 
 void    fanSetSpeed(uint8_t i, uint8_t speed);
 uint8_t fanGetSpeed(uint8_t i);
-void    fanSetSpeedPercent(uint8_t i, int16_t percent);
+void    fanSetSpeedPercent(uint8_t i, uint8_t percent);
 uint8_t fanGetSpeedPercent(uint8_t i);
 void    fanSetSendWaiting(uint8_t i, bool isWaiting);
